+++ qoqo/tests/integration/operations/multi_qubit_gate_operations.rs
@@ -18,7 +18,9 @@ use pyo3::Python;
 use qoqo::operations::convert_operation_to_pyobject;
 #[cfg(feature = "unstable_operation_definition")]
 use qoqo::operations::CallDefinedGateWrapper;
-use qoqo::operations::{MultiQubitCNOTWrapper, MultiQubitMSWrapper, MultiQubitZZWrapper};
+use qoqo::operations::{
+    MultiQubitCNOTWrapper, MultiQubitMSWrapper, MultiQubitZZWrapper, QFTWrapper,
+};
 use qoqo::CircuitWrapper;
 use qoqo_calculator::Calculator;
 use qoqo_calculator::CalculatorFloat;
@@ -28,8 +30,11 @@ use roqoqo::operations::*;
 #[cfg(feature = "json_schema")]
 use roqoqo::ROQOQO_VERSION;
 use roqoqo::{Circuit, RoqoqoError};
-use std::collections::HashMap;
 use std::convert::TryInto;
+use std::{
+    collections::HashMap,
+    f64::consts::{FRAC_PI_2, FRAC_PI_4},
+};
 use test_case::test_case;
 
 use super::convert_cf_to_pyobject;
@@ -148,6 +153,7 @@ fn test_new_call_defined_gate(
         );
     })
 }
+
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1])), vec![0, 1], "__eq__"; "MultiQubitCNOT_eq")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![2, 3])), vec![0, 1], "__ne__"; "MultiQubitCNOT_ne")]
 fn test_new_multi_cnot(input_operation: Operation, arguments: Vec<u32>, method: &str) {
@@ -184,6 +190,49 @@ fn test_new_multi_cnot(input_operation: Operation, arguments: Vec<u32>, method: 
     })
 }
 
+#[test_case(Operation::from(QFT::new(vec![0, 1], true, true)), (vec![0, 1], true, true), "__eq__"; "QFT_eq")]
+#[test_case(Operation::from(QFT::new(vec![2, 3], true, true)), (vec![0, 1], true, true), "__ne__"; "QFT_ne")]
+fn test_new_qft(input_operation: Operation, arguments: (Vec<u32>, bool, bool), method: &str) {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let operation = convert_operation_to_pyobject(input_operation, py).unwrap();
+        // Basic initialisation, no errors
+        let operation_type = py.get_type::<QFTWrapper>();
+        let binding = operation_type.call1(arguments).unwrap();
+        let operation_py = binding.downcast::<QFTWrapper>().unwrap();
+        let comparison =
+            bool::extract_bound(&operation.call_method1(method, (operation_py,)).unwrap()).unwrap();
+        assert!(comparison);
+
+        // Error initialisation
+        let result = operation_type.call1((vec!["fails"], false, false));
+        let result_ref = result.as_ref();
+        assert!(result_ref.is_err());
+
+        // Testing PartialEq, Clone and Debug
+        let def_wrapper = operation_py.extract::<QFTWrapper>().unwrap();
+        let binding = operation_type.call1((vec![1, 2], true, true)).unwrap();
+        let new_op_diff = binding.downcast::<QFTWrapper>().unwrap();
+        let def_wrapper_diff = new_op_diff.extract::<QFTWrapper>().unwrap();
+        let helper_ne: bool = def_wrapper_diff != def_wrapper;
+        assert!(helper_ne);
+        let helper_eq: bool = def_wrapper == def_wrapper.clone();
+        assert!(helper_eq);
+
+        assert_eq!(
+            format!("{def_wrapper_diff:?}"),
+            "QFTWrapper { internal: QFT { qubits: [1, 2], swaps: true, inverse: true } }"
+        );
+
+        // Testing inputs
+        let swaps_input = bool::extract_bound(&operation.call_method0("swaps").unwrap()).unwrap();
+        assert!(swaps_input);
+        let inverse_input =
+            bool::extract_bound(&operation.call_method0("inverse").unwrap()).unwrap();
+        assert!(inverse_input);
+    })
+}
+
 /// Test is_parametrized() function for MultiQubitGate Operations
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1], CalculatorFloat::from("theta"))); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1], CalculatorFloat::from("theta"))); "MultiQubitZZ")]
@@ -203,6 +252,7 @@ fn test_pyo3_is_parametrized(input_operation: Operation) {
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1], CalculatorFloat::PI)); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1], CalculatorFloat::PI)); "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1])); "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1], true, true)); "QFT")]
 fn test_pyo3_is_not_parametrized(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -240,6 +290,9 @@ fn test_pyo3_theta(theta: CalculatorFloat, input_operation: Operation) {
 #[test_case(vec![0], Operation::from(MultiQubitCNOT::new(vec![0])); "MultiQubitCNOT one")]
 #[test_case(vec![0, 1], Operation::from(MultiQubitCNOT::new(vec![0, 1])); "MultiQubitCNOT two")]
 #[test_case(vec![0, 1, 2], Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])); "MultiQubitCNOT three")]
+#[test_case(vec![0], Operation::from(QFT::new(vec![0], true, true)); "QFT one")]
+#[test_case(vec![0, 1], Operation::from(QFT::new(vec![0, 1], true, true)); "QFT two")]
+#[test_case(vec![0, 1, 2], Operation::from(QFT::new(vec![0, 1, 2], true, true)); "QFT three")]
 fn test_pyo3_qubits(qubit: Vec<usize>, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -253,6 +306,7 @@ fn test_pyo3_qubits(qubit: Vec<usize>, input_operation: Operation) {
 #[test_case("MultiQubitMS", Operation::from(MultiQubitMS::new(vec![0, 1], CalculatorFloat::from(0))); "MultiQubitMS")]
 #[test_case("MultiQubitZZ", Operation::from(MultiQubitZZ::new(vec![0, 1], CalculatorFloat::from(0))); "MultiQubitZZ")]
 #[test_case("MultiQubitCNOT", Operation::from(MultiQubitCNOT::new(vec![0, 1])); "MultiQubitCNOT")]
+#[test_case("QFT", Operation::from(QFT::new(vec![0, 1], true, false)); "QFT")]
 fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -294,6 +348,15 @@ fn test_pyo3_hqslang(name: &'static str, input_operation: Operation) {
         "MultiQubitCNOT",
         ];
     "MultiQubitCNOT")]
+#[test_case(
+    Operation::from(QFT::new(vec![0, 1, 2], false, true)),
+    vec![
+        "Operation",
+        "GateOperation",
+        "MultiQubitGateOperation",
+        "QFT",
+        ];
+    "QFT")]
 fn test_pyo3_tags(input_operation: Operation, tags: Vec<&str>) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -380,6 +443,7 @@ fn test_pyo3_call_defined_gate_inputs() {
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])); "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1, 2], true, true)); "QFT")]
 fn test_pyo3_remapqubits(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -441,6 +505,7 @@ fn test_pyo3_remapqubits_call_defined_gate() {
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])); "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1, 2], false, false)); "QFT")]
 fn test_pyo3_remapqubits_error(input_operation: Operation) {
     // preparation
     pyo3::prepare_freethreaded_python();
@@ -482,6 +547,7 @@ fn test_pyo3_remapqubits_error_call_defined_gate() {
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])); "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1, 2], true, false)); "QFT")]
 fn test_pyo3_unitarymatrix(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -563,6 +629,7 @@ fn test_pyo3_circuit_zz() {
         assert_eq!(result_circuit.internal, circuit);
     })
 }
+
 /// Test circuit() function for MultiQubitCNOT
 #[test]
 fn test_pyo3_circuit_multi_cnot() {
@@ -594,10 +661,33 @@ fn test_pyo3_circuit_multi_cnot() {
     })
 }
 
+/// Test circuit() function for QFT
+#[test]
+fn test_pyo3_circuit_qft() {
+    pyo3::prepare_freethreaded_python();
+    Python::with_gil(|py| {
+        let input_operation = Operation::from(QFT::new(vec![0, 1, 2], false, false));
+        let operation = convert_operation_to_pyobject(input_operation, py).unwrap();
+        let py_result = operation.call_method0("circuit").unwrap();
+        let result_circuit: CircuitWrapper = py_result.extract().unwrap();
+
+        let mut circuit = Circuit::new();
+        circuit += Hadamard::new(0);
+        circuit += ControlledPhaseShift::new(1, 0, FRAC_PI_2.into());
+        circuit += ControlledPhaseShift::new(2, 0, FRAC_PI_4.into());
+        circuit += Hadamard::new(1);
+        circuit += ControlledPhaseShift::new(2, 1, FRAC_PI_2.into());
+        circuit += Hadamard::new(2);
+
+        assert_eq!(result_circuit.internal, circuit);
+    })
+}
+
 /// Test copy and deepcopy functions
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1, 2], CalculatorFloat::from(1.3))); "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])); "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1, 2], true, true)); "QFT")]
 fn test_pyo3_copy_deepcopy(input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -672,6 +762,10 @@ fn test_pyo3_copy_deepcopy_call_defined_gate() {
     "MultiQubitCNOT { qubits: [0, 1, 2] }",
     Operation::from(MultiQubitCNOT::new(vec![0, 1, 2]));
     "MultiQubitCNOT")]
+#[test_case(
+    "QFT { qubits: [0, 1, 2], swaps: false, inverse: false }",
+    Operation::from(QFT::new(vec![0, 1, 2], false, false));
+    "QFT")]
 fn test_pyo3_format_repr(format_repr: &str, input_operation: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -853,6 +947,9 @@ fn test_pyo3_rotate_powercf(first_op: Operation, second_op: Operation) {
 #[test_case(
     Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])),
     Operation::from(MultiQubitCNOT::new(vec![1, 2])); "MultiQubitCNOT")]
+#[test_case(
+    Operation::from(QFT::new(vec![0, 1, 2], false, false)),
+    Operation::from(QFT::new(vec![0, 1, 2], true, false)); "QFT")]
 fn test_pyo3_richcmp(definition_1: Operation, definition_2: Operation) {
     pyo3::prepare_freethreaded_python();
     Python::with_gil(|py| {
@@ -936,6 +1033,7 @@ fn test_pyo3_richcmp_call_defined_gate() {
 #[test_case(Operation::from(MultiQubitMS::new(vec![0, 1, 2], CalculatorFloat::from(0))), "1.0.0"; "MultiQubitMS")]
 #[test_case(Operation::from(MultiQubitZZ::new(vec![0, 1, 2], CalculatorFloat::from(0))), "1.0.0"; "MultiQubitZZ")]
 #[test_case(Operation::from(MultiQubitCNOT::new(vec![0, 1, 2])), "1.20.0"; "MultiQubitCNOT")]
+#[test_case(Operation::from(QFT::new(vec![0, 1, 2], true, true)), "1.20.0"; "QFT")]
 fn test_pyo3_json_schema(operation: Operation, version_string: &str) {
     let rust_schema = match operation {
         Operation::MultiQubitMS(_) => {
@@ -947,6 +1045,7 @@ fn test_pyo3_json_schema(operation: Operation, version_string: &str) {
         Operation::MultiQubitCNOT(_) => {
             serde_json::to_string_pretty(&schemars::schema_for!(MultiQubitCNOT)).unwrap()
         }
+        Operation::QFT(_) => serde_json::to_string_pretty(&schemars::schema_for!(QFT)).unwrap(),
         _ => unreachable!(),
     };
     pyo3::prepare_freethreaded_python();